servo_tracker/raspberry_pi/requirements.txt
# Raspberry Pi Requirements
RPi.GPIO>=0.7.0
pyserial>=3.5
